+++ packages/html-elements/src/css/className.ts
@@ -1,0 +1,3 @@
+export function withClassName(styles: any[] | any, className?: string) {
+  return styles;
+}