+++ packages/html-elements/src/css/className.web.ts
@@ -1,0 +1,9 @@
+export function withClassName(styles: any[] | any, className?: string) {
+  if (className) {
+    if (!Array.isArray(styles)) {
+      styles = [styles];
+    }
+    return [...styles, { $$css: true, _: className }];
+  }
+  return styles;
+}