+++ packages/html-elements/src/css/createClassNameView.tsx
@@ -1,0 +1,5 @@
+import React from 'react';
+
+export function createClassNameView<TView extends React.ComponentType<any>>(View: TView) {
+  return View;
+}